+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Chg: Now using `Int16Array` instead of `Uint16Array` for voxel data
 - Chg: Introduced `ostrich.OstrichImageData` which is now taken by `ostrich.initializeElement()`
 - Fix: Fixed data-order to z-y-x
+- Fix: Linear VOI transform was off center
 - Add: Added tool `ostrich.VoiWindowingTool`
 - Add: Implemented Value Of Interest (VOI) windowing
 - Add: Added event `ostrich.voiWindowChanged`